+++ R/detect_separation.R
@@ -1,4 +1,4 @@
-# Copyright (C) 2017 Ioannis Kosmidis
+# Copyright (C) 2017-2019 Ioannis Kosmidis
 
 #  This program is free software; you can redistribute it and/or modify
 #  it under the terms of the GNU General Public License as published by
@@ -29,14 +29,14 @@
 #'
 #' @aliases detectSeparation print.detect_separation
 #'
-#' @param x \code{x} is a design matrix of dimension \code{n * p},
-#' @param y \code{y} is a vector of observations of length \code{n}
+#' @param x \code{x} is a design matrix of dimension \code{n * p}.
+#' @param y \code{y} is a vector of observations of length \code{n}.
 #' @param control a list of parameters controlling separation
 #'     detection. See \code{\link{detect_separation_control}} for
 #'     details.
-#' @param start currently not used
-#' @param mustart currently not used
-#' @param etastart currently not used
+#' @param start currently not used.
+#' @param mustart currently not used.
+#' @param etastart currently not used.
 #' @param singular.ok logical. If \code{FALSE}, a singular model is an
 #'     error.
 #' @param ... arguments to be used to form the default 'control'
@@ -65,12 +65,12 @@
 #'
 #' @references
 #'
-#' Kjell Konis (2007). *Linear Programming Algorithms for Detecting
+#' Kjell K. (2007). *Linear Programming Algorithms for Detecting
 #' Separated Data in Binary Logistic Regression
 #' Models*. DPhil. University of Oxford.
 #' \url{https://ora.ox.ac.uk/objects/uuid:8f9ee0d0-d78e-4101-9ab4-f9cbceed2a2a}
 #'
-#' Kjell Konis (2013). safeBinaryRegression: Safe Binary Regression. R
+#' Kjell K. (2013). safeBinaryRegression: Safe Binary Regression. R
 #' package version 0.1-3.
 #' \url{https://CRAN.R-project.org/package=safeBinaryRegression}
 #'
@@ -203,7 +203,7 @@ detect_separation <- function (x, y, weights = rep(1, nobs),
 #'     \code{"test"} for separation or also \code{"find"} which
 #'     parameters are infinite?
 #' @param beta_tolerance maximum absolute variable value from the
-#'     linear program, before separation is declared
+#'     linear program, before separation is declared.
 #'
 #' @export
 detect_separation_control <- function(linear_program = c("primal", "dual"),