+++ NEWS.md
@@ -1,5 +1,25 @@
+# brglm 0.5.0
+## Bug fixes
+* `brglmFit` now works as expected with custom link functions (mean and median bias reduction)
+* `brglmFit` respects the specification of the transformation argument in `brglmControl`
+* Fixed bug in the computation of the QR decomposition under aliasing in `brglmFit`
+* Other minor bug fixes and performance improvements
+* Protection against use of `quasi`, `quasibinomial` and `quasibinomial` families and documentation update.
+
+## New functionality
+* Added `bracl` for fitting adjacent category logit models for ordinal responses using maximum likelihood, mean bias reduction, and median bias reduction and associated methods (`logLik`, `summary` and so on)
+* Added `predict` methods for `brmultinom` and `bracl`
+@ Added `residuals` methods for `brmultinom` and `bracl` (residuals of the equivalent Poisson log-linear model)
+* Added the `mis` link functions for accounting for misclassification in binomial response models (Neuhaus, 1999, Biometrika)
+
+## Other improvements, updates and additions
+* Improved `summary` method for `brmultinom` objects
+* Better starting values for null fits
+* Added references to [arxiv:1804.04085](https://arxiv.org/abs/1804.04085) in documentation
+* Updated reference to [Kenne Pagui et al (2017)](https://doi.org/10.1093/biomet/asx046)
+
 # brglm 0.1.8
-## Other improvements, updates and addition
+## Other improvements, updates and additions
 * Improved documentation examples
 * Removed warning about observations with non-positive weights from brmultinom
 * Updated email adrress for Ioannis Kosmidis in brglmFit
@@ -10,7 +30,7 @@
 
 # brglm 0.1.7
 
-## Other improvements, updates and addition
+## Other improvements, updates and additions
 * Eliminated errors from markdown chunks in multinomial vignette
 
 # brglm2 0.1.6
@@ -18,9 +38,7 @@
 ## Bug fixes
 * Compatibility with new version of enrichwith
 
-## New functionality
-
-## Other improvements, updates and addition
+## Other improvements, updates and additions
 * New email for Ioannis Kosmidis
 
 # brglm2 0.1.5
@@ -32,7 +50,7 @@
 * `check_infinite_estimates` now accepts `brmultinom` objects
 * Added `singular.ok` argument to `brglmFit` and `detect_separation` methods in line with the update of `glm.fit`
 
-## Other improvements, updates and addition
+## Other improvements, updates and additions
 * less strict tolerance in `brglm_control`
 * Updates to help files
 * Fixed typos in iteration vignette